Remove v1 RangeDelAggregator
Summary: Now that v2 is fully functional, the v1 aggregator is removed.
The v2 aggregator has been renamed.

Test Plan: make check
