+++ TARGETS
@@ -124,7 +124,6 @@ cpp_library(
         "db/merge_helper.cc",
         "db/merge_operator.cc",
         "db/range_del_aggregator.cc",
-        "db/range_del_aggregator_v2.cc",
         "db/range_tombstone_fragmenter.cc",
         "db/repair.cc",
         "db/snapshot_impl.cc",
@@ -934,11 +933,6 @@ ROCKS_TESTS = [
     [
         "range_del_aggregator_test",
         "db/range_del_aggregator_test.cc",
-        "serial",
-    ],
-    [
-        "range_del_aggregator_v2_test",
-        "db/range_del_aggregator_v2_test.cc",
         "serial",
     ],
     [
