+++ utilities/debug.cc
@@ -19,8 +19,8 @@ Status GetAllKeyVersions(DB* db, Slice begin_key, Slice end_key,
 
   DBImpl* idb = static_cast<DBImpl*>(db->GetRootDB());
   auto icmp = InternalKeyComparator(idb->GetOptions().comparator);
-  ReadRangeDelAggregatorV2 range_del_agg(&icmp,
-                                         kMaxSequenceNumber /* upper_bound */);
+  ReadRangeDelAggregator range_del_agg(&icmp,
+                                       kMaxSequenceNumber /* upper_bound */);
   Arena arena;
   ScopedArenaIterator iter(
       idb->NewInternalIterator(&arena, &range_del_agg, kMaxSequenceNumber));
